datapath: Redirect nodeport L7 LB services
Redirect L7 LB nodeport service connections received by the host to an
Envoy listener. Prior to this commit these services worked as intended
only when the service connection was originated by a pod. This commit
implements L7 LB redirection for nodeport services from the local node,
or received by the node from other nodes or from an external load
balancer.

Redirect to nodeport L7 LB from bpf_sock for hostns sources. This is
needed due to this case not being handled by eBPF datapath on TC
level. Note that this requires a recent kernel to work (Linux
5.12+). With older kernels nodeport services only work between nodes,
not from a node host namespace to the node itself.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>