accesslog: Pass destination identity from L7 proxies to access logs
Pass destination security identity, as used at the time of policy
enforcement, from L7 proxies to access logs. This can be 0 if not
resolved by the proxy (e.g., by an ingress proxy), but in those cases the
logger now resolves the destination IP to the corresponding security
identity even when the destination is not a local endpoint.

Remove redundant code from pkg/fqdn and simplify pkg/proxy/logger now
that the source and destination endpoint info is filled in the same way.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>